+++ web-components/aofl-list-option/documentation/js/__config/rotation-conditions/index.js
@@ -1,0 +1,9 @@
+import routes from './routes';
+import ios from './ios';
+import monthlyTest from './monthly-test';
+
+export default {
+  routes,
+  ios,
+  'monthly-test': monthlyTest
+};