+++ web-components/aofl-list-option/documentation/routes.config.js
@@ -1,2 +1,13 @@
 window.aofljsConfig = window.aofljsConfig || {};
-window.aofljsConfig.routesConfig = {};
+window.aofljsConfig.routesConfig = {
+  'routes': [
+    {
+      'resolve': () => import('./routes/home-element/index.js'),
+      'rotation': 'routes',
+      'path': '/',
+      'dynamic': false,
+      'title': 'AofL::Home',
+      'locale': ''
+    }
+  ]
+};