+++ web-components/aofl-list-option/documentation/routes/home-element/examples/index.js
@@ -1,0 +1,3 @@
+context.toggleDrawer = () => {
+  context.drawer = !context.drawer;
+};