+++ web-components/aofl-list-option/documentation/routes/home-element/examples/template.js
@@ -1,0 +1,6 @@
+(context, html) => html`
+<aofl-select-list>
+  <aofl-list-option>Select me</aofl-list-option>
+  <aofl-list-option>Or me</aofl-list-option>
+</aofl-select-list>
+`;