+++ web-components/aofl-list-option/template.js
@@ -1,12 +1,8 @@
-import {html} from '@polymer/lit-element';
-
-export const template = (context) => html`
+export const template = (context, html) => html`
 <style>
 :host {
   display: block;
 }
 </style>
-<div on-click="${() => context.select()}">
-  <slot></slot>
-</div>
+<slot></slot>
 `;